pyproject.toml
[project]
name = "salary-predictor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "matplotlib>=3.10.8",
    "pandas>=3.0.2",
    "plotly>=6.0.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "scikit-learn>=1.8.0",
    "streamlit>=1.44.0",
    "supabase>=2.28.3",
    "uvicorn>=0.44.0",
]
